moved Modal.jsx to a separate folder, added Modal.scss and moved all styles to it
