+++ src/component/utils/Modal/Modal.scss
@@ -1,0 +1,54 @@
+.modal {
+  width: 60%;
+  height: 70%;
+
+  max-width: 600px;
+  max-height: 570px;
+
+  position: fixed;
+  top: 15%;
+  left: calc((100% - 600px) / 2);
+
+  padding: 16px;
+
+  border: 1px solid #ccc;
+  border-radius: 10px;
+
+  z-index: 20;
+
+  background-color: #f8f8f8;
+
+  box-shadow: 1px 1px 1px black;
+
+  transition: all 0.3s ease-out;
+
+  transform: translateY(-100vh);
+  opacity: 0;
+
+  overflow-y: scroll;
+
+  &--show {
+    transform: translateY(0);
+    opacity: 1;
+  }
+}
+
+.backdrop {
+  width: 100%;
+  height: 100%;
+
+  position: fixed;
+
+  left: 0;
+  top: 0;
+
+  background-color: rgba(0, 0, 0, 0.5);
+
+  z-index: 10;
+}
+
+@media (max-width: 416px) {
+  .modal {
+    left: 2%;
+  }
+}
